move heatmap components
